chore(site): Move Storybook Webpack config to a separate filei - Exclude external modules for babel and styles - Move Webpack config to it's own file and directory
